+++ wiki/Changelog.md
@@ -7,6 +7,7 @@
 * 增加参数 `checkExampleEntityClass`，默认 `false` 用于校验通用 Example 构造参数 entityClass 是否和当前调用的 Mapper<EntityClass> 类型一致 #201
 * 增加参数 `useSimpleType`，默认 `false`，启用后判断实体类属性是否为表字段时校验字段是否为简单类型，如果不是就忽略该属性，这个配置优先级高于所有注解
 * 增加参数 `simpleTypes`，默认的简单类型在 `SimpleTypeUtil` 中，使用该参数可以增加额外的简单类型，通过逗号隔开的全限定类名添加
+* 所有 `RuntimeException` 异常改为 `tk.mybatis.mapper.MapperException` 异常
 
 ##3.3.9 - 2016-09-04